requirements.txt
google-cloud-storage==2.19.0
python-dotenv==1.0.1